Move everything from folder internationaal to international (#3143)
* Move folder

* Remove mock data
