+++ packages/app/src/pages/internationaal/positief-geteste-mensen.tsx
@@ -12,8 +12,7 @@ import { PageInformationBlock } from '~/components/page-information-block';
 import { TileList } from '~/components/tile-list';
 import { TimeSeriesChart } from '~/components/time-series-chart';
 import { LineSeriesDefinition } from '~/components/time-series-chart/logic';
-import { EuropeChoroplethTile } from '~/domain/internationaal/europe-choropleth-tile';
-import { choroplethMockData } from '~/domain/internationaal/logic/choropleth-mock-data';
+import { EuropeChoroplethTile } from '~/domain/international/europe-choropleth-tile';
 import { InfectedTableTile } from '~/domain/international/infected-table-tile';
 import { CountryOption } from '~/domain/international/select-countries/context';
 import {
@@ -63,7 +62,7 @@ export const getStaticProps = withFeatureNotFoundPage(
     }`;
     }),
     createGetChoroplethData({
-      in: ({ tested_overall }) => tested_overall || choroplethMockData(),
+      in: ({ tested_overall }) => tested_overall,
     }),
     () => {
       const { internationalData } = getInData([...countryCodes])();
